Match start

+++ match/match.service.js
@@ -1,6 +1,8 @@
 const db = require('_helpers/db');
 const Match = db.Match;
 const users = require('users/user.service');
+const utils = require('_helpers/utils');
+
 
 module.exports = {
     getAll,
@@ -8,7 +10,8 @@ module.exports = {
     create,
     join,
     leave,
-    delete: _delete
+    delete: _delete,
+    start
 };
 
 async function getAll() {
@@ -38,6 +41,8 @@ async function join(id, matchParam) {
     const match = await Match.findById(id);
 
     if (!match) throw 'Match not found';
+
+    if (match.stage !== "CREATED") throw 'Match already started'
 
     if (match.players.size() = match.size) throw 'Match full';
 
@@ -73,6 +78,8 @@ async function leave(id, matchParam) {
 
     if (!match) throw 'Match not found';
 
+    if (match.stage !== "CREATED") throw 'Match already started'
+
     if (match.players.size() = match.size) throw 'Match full';
 
     const user = await users.getById(matchParam.userToRemove);
@@ -89,5 +96,57 @@ async function leave(id, matchParam) {
 }
 
 async function _delete(id) {
+    const match = await Match.findById(id);
+
+    if (!match) throw 'Wrong match id'
+
+    if (match.stage !== "CREATED") throw 'Match already started'
+
     await Match.findByIdAndRemove(id);
+}
+
+async function start(id) {
+    const match = await Match.findById(id);
+
+    if (!match) throw 'Match not found';
+
+    if (match.stage !== "CREATED") throw 'Match already started'
+
+    match.stage = "STARTED";
+
+    const countriestemp = Object.entries(match.countries);
+    const countriestemp2 = Object.entries(match.countries);
+
+    const promedio = Math.trunc(countriestemp.size()/match.players.size());
+
+    match.players.forEach(p => {
+        for (let i = 0; i < promedio; i++) {
+            const num = utils.getRndInteger(0,countriestemp.size()-1);
+            const name = countriestemp[num][0]
+            const where = countriestemp2.findIndex(c => c[0] === name);
+            countriestemp2[where][1].owner = p.username;
+            countriestemp2[where][1].armies = 3;
+            countriestemp.splice(num,1);
+        }
+    });
+
+    match.players.forEach(p => {
+        if (countriestemp.size()>0) {
+            const name = countriestemp[0][0]
+            const where = countriestemp2.findIndex(c => c[0] === name);
+            countriestemp2[where][1].owner = p.username;
+            countriestemp2[where][1].armies = 3;
+            countriestemp.splice(num,1);
+        } else {
+            break;
+        }
+    })
+
+    match.countries = Object.fromEntries(countriestemp2);
+
+    match.currentPlayer = match.players[utils.getRndInteger(0,countriestemp2.size()-1)].username;
+
+    await match.save();
+
+    return match;
 }
